Remove duplicate Android Connector class
