--- workers/unity/Assets/Fps/Scripts/SetupLogic/AndroidClientWorkerConnector.cs
@@ -1,93 +1,0 @@
-﻿using System;
-using Improbable.Gdk.Core;
-using Improbable.Gdk.Mobile;
-using UnityEngine;
-#if UNITY_ANDROID
-using Improbable.Gdk.Mobile.Android;
-
-#endif
-
-
-namespace Fps
-{
-    public class AndroidClientWorkerConnector : MobileWorkerConnector, IMobileConnectionController
-    {
-        public string IpAddress { get; set; }
-        public ConnectionScreenController ConnectionScreenController { get; set; }
-
-        [SerializeField] private GameObject level;
-
-        private GameObject levelInstance;
-
-        public async void TryConnect()
-        {
-            await Connect(WorkerUtils.AndroidClient, new ForwardingDispatcher()).ConfigureAwait(false);
-        }
-
-        protected override void HandleWorkerConnectionEstablished()
-        {
-            ConnectionScreenController.OnConnectionSucceeded();
-            WorkerUtils.AddClientSystems(Worker.World);
-
-            if (level == null)
-            {
-                return;
-            }
-
-            levelInstance = Instantiate(level, transform.position, transform.rotation);
-        }
-
-        protected override void HandleWorkerConnectionFailure(string errorMessage)
-        {
-            ConnectionScreenController.OnConnectionFailed(errorMessage);
-        }
-
-        protected override string GetHostIp()
-        {
-#if UNITY_ANDROID
-            if (!string.IsNullOrEmpty(IpAddress))
-            {
-                return IpAddress;
-            }
-
-            if (Application.isMobilePlatform && AndroidDeviceInfo.ActiveDeviceType == MobileDeviceType.Virtual)
-            {
-                return AndroidDeviceInfo.EmulatorDefaultCallbackIp;
-            }
-
-            return RuntimeConfigDefaults.ReceptionistHost;
-#else
-            throw new PlatformNotSupportedException(
-                $"{nameof(AndroidClientWorkerConnector)} can only be used for the Android platform. Please check your build settings.");
-#endif
-        }
-
-        public override void Dispose()
-        {
-            if (levelInstance != null)
-            {
-                Destroy(levelInstance);
-            }
-
-            base.Dispose();
-        }
-
-        public void UseIpAddressFromArguments()
-        {
-            IpAddress = GetReceptionistHostFromArguments();
-
-            if (string.IsNullOrEmpty(IpAddress))
-            {
-                IpAddress = "127.0.0.1";
-            }
-        }
-
-        private string GetReceptionistHostFromArguments()
-        {
-            var arguments = LaunchArguments.GetArguments();
-            var hostIp =
-                CommandLineUtility.GetCommandLineValue(arguments, RuntimeConfigNames.ReceptionistHost, string.Empty);
-            return hostIp;
-        }
-    }
-}
