--- workers/unity/Assets/Fps/Scripts/SetupLogic/AndroidClientWorkerConnector.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 40d797d47254a9649b733da47bb3aab8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
